minipc+single shoot+BT gimbal ctrl

+++ Core/Src/Gimbal_Control/gim_shoot_ctrl.c
@@ -5,7 +5,7 @@
  * @Author       : GDDG08
  * @Date         : 2021-12-22 22:06:02
  * @LastEditors  : GDDG08
- * @LastEditTime : 2022-06-22 20:34:07
+ * @LastEditTime : 2022-06-23 17:38:31
  */
 
 #include "gim_shoot_ctrl.h"
@@ -316,7 +316,7 @@ uint8_t Shooter_GetRefereeOverSpeed() {
 
 float Shooter_GetRefereeSpeedFdb(void) {
     BusComm_BusCommDataTypeDef* buscomm = BusComm_GetBusDataPtr();
-    uint16_t shooter_speed = buscomm->speed_17mm_fdb;
+    float shooter_speed = buscomm->speed_17mm_fdb;
     if (!(shooter_speed > 0) && (shooter_speed < 35)) {
         switch (buscomm->speed_17mm_limit) {
             case REFEREE_SHOOTER_SPEED_15:
